+++ PortClient_2/src/info/Point3D.java
@@ -5,35 +5,35 @@ import java.io.Serializable;
 public class Point3D implements Serializable{
 	private static final long serialVersionUID = -1666335444287045879L;
 	
-	private double x, y, z;
+	private float x, y, z;
 	
-	public Point3D(double x, double y, double z){
+	public Point3D(float x, float y, float z){
 		this.setX(x);
 		this.setY(y);
 		this.setZ(z);
 	}
 
-	public double getX() {
+	public float getX() {
 		return x;
 	}
 
-	public void setX(double x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public double getY() {
+	public float getY() {
 		return y;
 	}
 
-	public void setY(double y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 
-	public double getZ() {
+	public float getZ() {
 		return z;
 	}
 
-	public void setZ(double z) {
+	public void setZ(float z) {
 		this.z = z;
 	}
